requirements.txt
django==3.1.3
django-model-utils==4.1.0
djangorestframework==3.12.2
markdown==3.3.4 
django-filter==2.4.0
psycopg2==2.8.6
fontawesome-free==5.15.2
django-utils-six==2.0